requirements.txt
pip==25.0.1
robotframework==7.1.1
robotframework-imagehorizonlibrary==1.0
robotframework-pabot==2.18.0
robotframework-seleniumlibrary==6.3.0
robotframework-stacktrace==0.4.1
robotframework-requests==0.9.7
robotframework-jsonschemalibrary==1.0
pycryptodomex~=3.22.0
crypto==1.4.1
browser==19.3.1
pytesseract==0.3.13
robotframework-jsonlibrary==0.5
numpy~=2.2.5
robotframework-tidy==4.16.0

opencv-python~=4.11.0.86
requests~=2.32.3
pillow~=11.2.1
pandas~=2.2.3